pyproject.toml
[project]
name = "html2md-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "loguru>=0.7.3",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
